Added support for ES6 modules to jest. Moved functions to helper.js and import them into the jest tests
