+++ selenium-with-jest/package.json
@@ -12,6 +12,8 @@
     "selenium-webdriver": "^4.0.0-alpha.7"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.12.1",
+    "babel-jest": "^26.6.3",
     "jest": "^26.6.3"
   }
 }
